requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
openai==1.54.4
httpx==0.27.2
python-multipart==0.0.6
werkzeug==2.3.7
eventlet==0.33.3
gunicorn==21.2.0
python-engineio==4.7.1
python-socketio==5.9.0
PyPDF2==3.0.1
python-docx==0.8.11
striprtf==0.0.26
pytesseract==0.3.10
Pillow==10.0.1
ocrmypdf>=16.0.0,<17.0.0
reportlab==4.0.4
aiohttp==3.9.1
beautifulsoup4==4.12.2
langchain==0.1.0
lxml==4.9.3
Brotli==1.1.0
pandas==2.2.3
